Modularizing one layer up, spreading dynamic testing to all rules added in this batch
