+++ test/rules/resources/cfnnag/test_RequiredResourceProperties.py
@@ -1,0 +1,12 @@
+from qs_cfn_lint_rules.RequiredResourceProperties import *
+from ... import BaseRuleTestCase, DynamicRuleTesting
+
+
+class DynamicRequiredResourceProperties(DynamicRuleTesting, BaseRuleTestCase):
+    imported_module = 'qs_cfn_lint_rules.RequiredResourceProperties'
+    module_attr = 'property_name'
+
+    def test_file_negative(self):
+        for cn, to in self.testobjs.items():
+            with self.subTest(f"Testing {cn} file negative"):
+                to.test_file_negative(2)
